+++ src/components/QuizError/QuizError.jsx
