+++ src/components/QuizError/QuizError.scss
